Refactor: modularize video browsing and chart control state logic into custom hooks
