+++ frontend_react/src/hooks/useChartControlState.ts
@@ -1,0 +1,16 @@
+import { useState } from "react";
+
+export type ChartType = "pie" | "bar";
+export type DurationUnit = "hours" | "minutes";
+
+export const useChartControlState = () => {
+  const [chartType, setChartType] = useState<ChartType>("pie");
+  const [durationUnit, setDurationUnit] = useState<DurationUnit>("hours");
+
+  return {
+    chartType,
+    setChartType,
+    durationUnit,
+    setDurationUnit,
+  };
+};
